refs #78, cleanup

+++ src/noit_defines.h
@@ -73,6 +73,8 @@ static inline void uuid_unparse_lower(uuid_t in, char *out) {
 /* udns checks for IPv6, noit doesn't work without it */
 #define HAVE_IPv6
 
+#include <stdio.h>
+
 static inline int noit_build_version(char *buff, int len) {
   const char *v = NOIT_HEADURL;
   const char *start, *end, *ns;
@@ -84,9 +86,9 @@ static inline int noit_build_version(char *buff, int len) {
       if(!strncmp(start, "trunk/", 6))
         return snprintf(buff, len, "trunk.%s", NOIT_SVNVERSION);
       if(!strncmp(start, "branches/", 9))
-        return snprintf(buff, len, "b_%.*s.%s", end - ns, ns, NOIT_SVNVERSION);
+        return snprintf(buff, len, "b_%.*s.%s", (int)(end - ns), ns, NOIT_SVNVERSION);
       if(!strncmp(start, "tags/", 5))
-        return snprintf(buff, len, "%.*s.%s", end - ns, ns, NOIT_SVNVERSION);
+        return snprintf(buff, len, "%.*s.%s", (int)(end - ns), ns, NOIT_SVNVERSION);
     }
   }
   return snprintf(buff, len, "unknown.%s", NOIT_SVNVERSION);
